Move constituency_id setting to callback
By calling store_constituency explicitly in the controller we
run the validations twice when creating the signature during
the create action.

By moving the setting of the constituency_id to a before_create
callback we can eliminate five database queries - the BEGIN/COMMIT
statements, two queries that check for duplicate signatures and
the UPDATE statement.