Move constituency lookup to signature validation
Since there's nothing to rate-limit requests to the signature creation
url there's a theoretical possibility that someone could DDoS the
service since that endpoint calls the Parliament constituency API and
is waiting on IO for approximately 80ms on average blocking the thread
from processing requests.

By moving the constituency lookup to the validation phase we do add
extra load to something that's already doing a lot of work but it's
effectively rate limited by our anti-fraud measures and isn't as
susceptible to bursts of requests since job queue helps to smooth that.

The lookup happens outside of the `retry_lock` block to ensure that
we're not adding 80ms to the length of time we're holding the lock
on the signature row.